+++ exercises.html
@@ -46,6 +46,6 @@
 
       </div>
   </div>
-  <script src="lib/js/exercises.js"></script>
+  <script src="main.bundle.js"></script>
 </body>
 </html>
